pyproject.toml
[project]
name = "european_values"
version = "0.0.0"
description = "Measuring European values of LLMs."
readme = "README.md"
authors = [
    {name = "Dan Saattrup Nielsen", email = "dan.nielsen@alexandra.dk"},
]
maintainers = [
    {name = "Dan Saattrup Nielsen", email = "dan.nielsen@alexandra.dk"},
]
requires-python = ">=3.11,<4.0"
dependencies = [
    "datasets>=3.6.0",
    "hydra-core>=1.3.2",
    "pandas>=2.2.3",
    "plotly>=6.1.0",
    "tqdm>=4.67.1",
    "umap-learn>=0.5.7",
]

[project.urls]
Repository = "https://github.com/alexandrainst/european_values"
Issues = "https://github.com/alexandrainst/european_values/issues"

[project.license]
file = "LICENSE"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "pre-commit>=3.8.0",
    "readme-coverage-badger>=0.1.2",
    "click>=8.1.7",
    "ruff>=0.11.1",
    "mypy>=1.15.0",
    "nbstripout>=0.8.1",
    "coverage>=5.5",
    "mkdocs-material>=9.5.45",
    "mkdocs-include-markdown-plugin>=7.0.1",
    "mkdocs-include-dir-to-nav>=1.2.0",
    "mkapi>=3.0.22",
    "pip>=24.3.1",
]

[tool.ruff]
target-version = "py311"
line-length = 88
exclude = [
    ".git",
    ".mypy_cache",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
]
extend-include = [
    "*.ipynb",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
skip-magic-trailing-comma = true

[tool.ruff.lint]
select = [
    # isort (imports)
    "I",
    # pydocstyle (docstrings)
    "D",
    # pycodestyle errors and warnings (PEP8)
    "E",
    "W",
    # flake-annotations (type hints)
    "ANN",
    # Pyflakes
    "F",
]
ignore = [
    # Type annotations for "self" arguments
    "ANN101",
    # Type annotations for "cls" arguments
    "ANN102",
    # Type annotations for **kwargs
    "ANN003",
    # Docstrings for **kwargs
    "D417",
]

[tool.ruff.lint.extend-per-file-ignores]
"__init__.py" = [
    "F401",
]
"src/scripts/main.py" = [
    "I",
]

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
    '--durations=10',
    '--color=yes',
    '--doctest-modules',
    '--cov=src/european_values',
]
xfail_strict = true
filterwarnings = [
    "error",
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore::ImportWarning",
    "ignore::ResourceWarning",
    "ignore::FutureWarning",
]
log_cli_level = "info"
testpaths = [
    "tests",
    "src/european_values",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
